Only flush event queue if event triggers a run
Fixes https://github.com/lpil/mix-test.watch/issues/65

On OSX the first events recieved when saving a file may not be attributed
to the Elixir file being edited, so they do not trigger a run. If we
then flush the watcher message queue the events that would trigger the
test run is discarded before being seen, causing the tests not to run.