Add reference to mix_test_interactive
Closes #115 

As suggested in https://github.com/lpil/mix-test.watch/issues/115#issuecomment-759513881, this adds a reference to the new [mix_test_interactive](https://github.com/influxdata/mix_test_interactive) project as an alternative.

Thanks again for this project! I've used it so much, and it served as a great base for `mix_test_interactive`!