Allow MIX_ENV/MIX_TARGET settings to pass through
Prior to this change, `MIX_ENV=test` was forced and any
`preferred_cli_env` and `preferred_cli_target` configuration in the
`mix.exs` would look like it was ignored. In actuality,
`preferred_cli_*` wasn't ignored, but you'd have to set it on `do` based
on how ANSI color was enabled.

This change enables ANSI a different way so that the mix target is
passed through and can get the `preferred_cli_*` settings, if any.
Because of this `MIX_ENV` doesn't need to be forced any more.