Change default exclusion regexes to binaries to support OTP 28
OTP 28 changes the implementation for regexes, and as a result you
can no longer use Regexes in module attributes.

I changed the defaults to strings of the patterns and added a function
clause to first compile the regexes if they are binary. This means
it is backwards compatible and the user can still configure actual
Regexes instead of binary patterns.